[simutil_verilator] Add public access to TraceOn/TraceOff
The methods used for starting and stopping tracing were previously private
to the VerilatorSimCtrl class but it's quite nice to have them public so
that they can be called via VerilatorSimCtrl::GetInstance().TraceOn().
This allows tracing to be started/stopped by C code running on the host
writing to DPI mailbox functions that call the relevant TraceOn/TraceOff.

Signed-off-by: Shareef Jalloq <shareef.jalloq@idexbiometrics.com>